backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "automata-lib>=9.0.0",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.10",
    "uvicorn>=0.34.2",
]
